added config_pack, all files compiled
A config package is added to the project. Two old memory controller related port type records (dsa_top_mem_con_type and mem_con_type are merged into a single mem_con_type with 6 signals which are enough to communicated with the AXI interface. The memcon type stuff are now in config_pack. Besides that, the configuration of system such as the reset value and memory range settings are moved to config_pack from malloc_pack.