in tb, sorted cmd vs we
as in simulation there isn't a requirement of using an enable signal for reading, however my logic does produce this signal. so I cannot connect start signal directly to we. sometimes it means read-enable. I added a small part of logic that helps separate write-enable and read-enable from the same m_request.start in tb_v0.