Modified interfaces and tb
Redundant signals and ports are removed such as total_entry who's replaced with a constant in configuration package. Other changes regarding ports and signals include the separation of the DSA done bit and the value returned. Testbench is modified in order to test the hash table accelerator. What needs to be done now: modify the test-generating C file.